requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
codecov==2.1.12
colorama==0.4.5
coverage==6.4.4
cycler==0.11.0
fonttools==4.37.1
frozenlist==1.3.1
idna==3.3
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.5.3
multidict==6.0.2
numpy==1.23.2
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
prettytable==3.4.1
PTable==0.9.2
py==1.11.0
pyaes==1.6.1
pycares==4.2.2
pycparser==2.21
pyparsing==3.0.9
Pyrogram==2.0.50
PySocks==1.7.1
pytest==7.1.3
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
pytest-cov==3.0.0
pytest-mock==3.8.2
python-dateutil==2.8.2
python-decouple==3.6
requests==2.28.1
six==1.16.0
TgCrypto==1.2.3
tomli==2.0.1
urllib3==1.26.12
wcwidth==0.2.5
yarl==1.8.1
